split get_batch from get.rs; add aqry_batch
